*** empty log message ***
git-svn-id: http://svn.vuze.com/public/client@1783 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/util/DisplayFormatters.java
@@ -179,9 +179,11 @@ DisplayFormatters
 			
 	  	if(pm != null){
 	  	
-			int nbFails = pm.getNbHashFails();
+			long nbFails = pm.getNbHashFails();
 		
-			long size = nbFails * dm.getPieceLength();
+				// size can exceed int so ensure longs used in multiplication
+			
+			long size = nbFails * (long)dm.getPieceLength();
 			
 			String result = nbFails + " ( ~ " + formatByteCountToKiBEtc(size) + " )";
